move common methods from util.util to util.op
